Simple Weather-App
The Simple Weather-App is designed to allow users to look up current weather conditions which include temperature, humidity and wind speed. It provides the user with an extensive list of cities that can be searched.

Jeremy K - https://github.com/jkunins95

Fork this repository
Clone the forked repository
Add your contributions (code or documentation)
Commit and push
Wait for pull request to be merged